修改角色禁用或删除后，缓存未更新

+++ rundo-auth-rbac/src/main/java/com/runjian/rbac/service/rbac/impl/UserServiceImpl.java
@@ -183,8 +183,10 @@ public class UserServiceImpl implements UserService {
                 existRoleIds.removeAll(difference);
                 // 移除交集获得增加的数据
                 roleIds.removeAll(difference);
-                // 删除去除的角色
-                userRoleMapper.deleteAllByUserIdAndRoleIds(userId, existRoleIds);
+                if (!existRoleIds.isEmpty()){
+                    // 删除去除的角色
+                    userRoleMapper.deleteAllByUserIdAndRoleIds(userId, existRoleIds);
+                }
             }
             // 保存新的角色
             if (!roleIds.isEmpty()) {
